Небольшой рефакторинг согласно критериям.
