+++ project/src/components/place-card/place-card-list/place-card-list.jsx
@@ -1,0 +1,11 @@
+import { PlaceCardListСities } from './place-card-list-cities/place-card-list-cities';
+import { PlaceCardListFavorites } from './place-card-list-favorites/place-card-list-favorites';
+import { PlaceCardListNearPlaces } from './place-card-list-near-places/place-card-list-near-places';
+
+const PlaceCardList = {
+  PlaceCardListСities,
+  PlaceCardListFavorites,
+  PlaceCardListNearPlaces,
+};
+
+export { PlaceCardList };
