+++ project/src/components/shared/logo/logo-header/logo-header.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import { Link } from 'react-router-dom';
+
+import { AppRoute } from '../../../../const';
+
+function LogoHeander() {
+  return (
+    <Link className="header__logo-link" to={AppRoute.ROOT}>
+      <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41" />
+    </Link>
+  );
+}
+
+export { LogoHeander };
