--- project/src/components/shared/logo/logo.jsx
@@ -1,22 +1,0 @@
-import React from 'react';
-import { Link } from 'react-router-dom';
-
-import { AppRoute } from '../../../const';
-
-function LogoHeander() {
-  return (
-    <Link className="header__logo-link" to={AppRoute.ROOT}>
-      <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41" />
-    </Link>
-  );
-}
-
-function LogoFooter() {
-  return (
-    <Link className="footer__logo-link" to={AppRoute.ROOT}>
-      <img className="footer__logo" src="img/logo.svg" alt="6 cities logo" width="64" height="33" />
-    </Link>
-  );
-}
-
-export { LogoHeander, LogoFooter };
